# config/swagger.yml
openapi: 3.0.0
info:
  title: Easy Procedures API
  description: REST API for managing bank procedures and requests.
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /users/login:
    post:
      tags:
        - Users
      summary: Authenticate and get a JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: secret
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      user:
                        type: object
                        properties:
                          id: {type: integer}
                          email: {type: string}
                          name: {type: string}
                          role_id: {type: integer}
        '401':
          description: Invalid credentials
  /users/register:
    post:
      tags:
        - Users
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - surname
                - email
                - phonenumber
                - password
                - confirm-password
              properties:
                name: {type: string, example: John}
                surname: {type: string, example: Doe}
                email: {type: string, format: email, example: john@example.com}
                phonenumber: {type: integer, example: 123456789}
                password: {type: string, format: password, example: secret}
                confirm-password: {type: string, format: password, example: secret}
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string, example: success}
                  message: {type: string, example: Registration successful.}
        '400':
          description: Validation error
  /users/forget-password:
    post:
      tags:
        - Users
      summary: Request a password reset link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email: {type: string, format: email, example: user@example.com}
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string, example: success}
                  message: {type: string, example: Instructions sent to your email.}
        '404':
          description: Email not found
  /procedures:
    get:
      tags:
        - Procedures
      summary: List all active procedures
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Procedure'
  /procedures/{id}:
    get:
      tags:
        - Procedures
      summary: Get a procedure by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Procedure'
        '404':
          description: Procedure not found
  /requests:
    get:
      tags:
        - Requests
      summary: List all requests for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Request'
    post:
      tags:
        - Requests
      summary: Start a new procedure request
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - procedure_id
              properties:
                procedure_id:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Request'
        '400':
          description: Validation error
  /requests/{id}:
    get:
      tags:
        - Requests
      summary: Get a request by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Request'
        '404':
          description: Request not found
  /requirements:
    get:
      tags:
        - Requirements
      summary: List all requirements
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Requirement'
  /requirements/{id}:
    get:
      tags:
        - Requirements
      summary: Get a requirement by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Requirement'
        '404':
          description: Requirement not found
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Procedure:
      type: object
      properties:
        id: {type: integer}
        name: {type: string}
        type: {type: string}
        description: {type: string}
        image: {type: string, nullable: true}
    Request:
      type: object
      properties:
        id: {type: integer}
        status: {type: string}
        user_id: {type: integer}
        procedure_id: {type: integer}
        created: {type: string, format: date-time}
    Requirement:
      type: object
      properties:
        id: {type: integer}
        name: {type: string}
        description: {type: string}
        status: {type: string}
        requirementtype_id: {type: integer}
security:
  - BearerAuth: []